refactor: simplify and convert to ts
