+++ client/src/features/viewers/public/Public.scss
@@ -5,149 +5,113 @@
   box-sizing: border-box; /* reset */
   overflow: hidden;
   width: 100%; /* restrict the page width to viewport */
-  background: var(--background-color-override, $viewer-background-color);
   height: 100vh;
+
+  font-family: $viewer-font-family;
+  background: var(--background-color-override, $viewer-background-color);
   color: var(--color-override, $viewer-color);
-  font-weight: 300;
+  gap: min(2vh, 16px);
+  padding: min(2vh, 16px) clamp(16px, 10vw, 64px);
 
   display: grid;
-  grid-template-columns: 1fr 1fr 1fr 3vw 2fr;
-  grid-template-rows: 15vh 1fr 1fr 0 13vh 13vh;
+  grid-template-columns: 1fr 1fr 40vw;
+  grid-template-rows: auto 12px 1fr auto;
   grid-template-areas:
-    ' titl titl titl . schd'
-    ' now  now  now  . schd'
-    ' next next .... . schd'
-    ' ...  .... .... . ....'
-    ' publ publ .... . info'
-    ' publ publ time . info';
-  gap: 1vw;
-  padding: 1vw;
+    ' header    header    header'
+    ' progress  progress  schedule-nav'
+    ' now       now       schedule'
+    ' message   message   info';
 
-  .event,
-  .public-container,
-  .info,
-  .timer-container,
-  .clock-container,
-  .today-container {
-    background-color: var(--outdent-background-color-override, $viewer-outdent-bg-color);
-    padding: 1vh 1vw;
-    border-radius: $element-border-radius;
+  /* =================== HEADER + EXTRAS ===================*/
 
-    .label {
-      font-size: 1.3vw;
-      color: var(--accent-color-override, $accent-color);
-    }
-
-    .clock,
-    .timer {
-      font-family: 'Open Sans', sans-serif;
-      font-size: 2.25vw;
-      line-height: 3vw;
-      text-align: center;
-      letter-spacing: 0.25vw;
-      color: var(--color-override, $viewer-color);
-    }
-  }
-
-  .event {
-    border-radius: 0 $element-border-radius $element-border-radius 0;
-    margin-left: -1vw;
-    padding: 1vh 2vw;
-    overflow: hidden;
-
-    &.now {
-      grid-area: now;
-    }
-    &.next {
-      grid-area: next;
-    }
-  }
-
-  .event-title {
-    grid-area: titl;
-    font-size: 3vw;
+  .event-header {
+    grid-area: header;
+    font-size: clamp(32px, 4.5vw, 64px);
     font-weight: 600;
-    text-decoration: underline 0.5vh;
-    text-decoration-color: var(--accent-color-override, $accent-color);
-    padding-top: 0.2vh;
-    padding-left: 1vw;
-  }
-
-  .today-container {
-    grid-area: schd;
-    padding: 2.5vh 2vw;
-    overflow: hidden;
-    margin-top: 3vh;
-    height: 95%;
-
-    .today-header-block {
-      display: flex;
-      justify-content: space-between;
-      margin-bottom: 1.5vh;
-
-      .nav {
-        align-self: center;
-        justify-content: flex-end;
-        display: flex;
-        margin-bottom: 2.5vh;
-      }
-
-      .nav-item {
-        background-color: $today-item-bg;
-        width: 0.7vw;
-        height: 0.7vw;
-        border-radius: 0.35vw;
-        margin-left: 0.5vw;
-      }
-    }
-  }
-
-  .public-container {
-    grid-area: publ;
-
-    &--hidden {
-      opacity: 0;
-      transition: 0.5s;
-      transition-property: opacity;
-    }
-
-    .message {
-      font-size: 1.5vw;
-      line-height: 2vw;
-      white-space: pre-line;
-      overflow: hidden;
-    }
+    display: flex;
   }
 
   .clock-container {
-    grid-area: time;
-    border-radius: $element-border-radius;
+    margin-left: auto;
+  }
+
+  .public-container {
+    grid-area: message;
+
+    &--hidden {
+      opacity: 0;
+      transition: $viewer-transition-time;
+      transition-property: opacity;
+    }
+  }
+
+  .clock-container,
+  .public-container {
+    .label {
+      font-size: clamp(16px, 1.5vw, 24px);
+      font-weight: 600;
+      color: var(--label-color-override, $viewer-label-color);
+      text-transform: uppercase;
+    }
+
+    .time {
+      font-size: clamp(32px, 3.5vw, 50px);
+      font-weight: 600;
+      letter-spacing: 0.05em;
+    }
+
+    .message {
+      font-size: clamp(16px, 1.5vw, 24px);
+      font-weight: 400;
+    }
+  }
+
+  /* =================== MAIN - NOW ===================*/
+
+  .now-container {
+    grid-area: now;
+    display: flex;
+    flex-direction: column;
+    gap: min(2vh, 16px);
+  }
+
+  .event {
+    background-color: var(--card-background-color-override, $viewer-card-bg-color);
+    padding: 16px 24px;
+    border-radius: 8px;
+  }
+
+  /* =================== MAIN - SCHEDULE ===================*/
+
+  .schedule-container {
+    grid-area: schedule;
+    overflow: hidden;
+    height: 100%;
+    margin-left: 16px;
+  }
+
+  .schedule-nav-container {
+    grid-area: schedule-nav;
+    align-self: center;
   }
 
   .info {
     grid-area: info;
-    display: grid;
-    padding: 2.5vh 2vw;
-
-    grid-template-rows: 3vh minmax(0, 1fr);
-    grid-template-columns: 3fr 1fr;
-    grid-template-areas:
-    'titl .'
-    'binf qr';
-    gap: 0.5vw;
+    display: flex;
+    gap: max(1vw, 16px);
 
     &__message {
-      grid-area: binf;
-      font-size: 1.5vw;
-      line-height: 2vw;
+      font-size: clamp(16px, 1.5vw, 24px);
+      line-height: 1.3em;
       white-space: pre-line;
       overflow: hidden;
+      flex: 1;
     }
 
     .qr {
-      align-self: center;
-      justify-self: center;
-      grid-area: qr;
+      margin-left: auto;
+      padding: 4px;
+      background-color: white;
     }
   }
 }
